Update APMg3 and tuning configs to version 9
This CL updates the APMg3 tunings to version 9. The changes performed is to reduce the post-AEC gain to the level that was in the v4 tunings (a 7 dB reduction), as the gain in v8 has been shown to sometimes saturate the nearend.

The tunings are available at
https://drive.google.com/drive/folders/11dKj-dpkNUlcci8kTsxi3lN5Ht5FEzXu?usp=drive_link
Complexity and memory figures are basically the same (slightly lower) as for https://docs.google.com/spreadsheets/d/1ZtPoIw3SGPKErR8xURvp3gyDvNNl7bISCzLFH_HDOIc/edit?usp=sharing&resourcekey=0-q7s7VtDqDSf762Lvb4d7gA

Bug: 302471415
Test: Manual test calls on Tangor using Meet. Lab testing in TPE labs with ACQUA/Headacoustics. Offline testing on recordings.
Change-Id: I12309ab01772724cda9da6bb793586bf0d79715e