Remove unnecessary Android.mk in device/google/tangorpro
After verification, the top-level Android.mk only includes the
Android.mk of its subdirectories. The only meaningful build target
in the subdirectory is `UwbVendorService`, but it is blocked by the
blank Android.mk in the upper layer and can never be built.
Therefore, all Android.mk in device/google/tangorpro are not necessary
to exist in reality.

Bug: 311511817
Test: N/A
Change-Id: I3c47169481cedc6a23e94a0d3d96bd2a644abe2a