Update APMg3 and tuning configs to version 11
This CL updates the APMg3 tunings to version 11.
The changes performed is to remove any signal level gains applied in APMg3 in AOC
for the headset usecase when the AEC is off. This eliminates the issue of 25 dB
gain being applied when headsets are used causing some headset audio signals to be
saturated.

This is a correction of the CL
https://googleplex-android-review.git.corp.google.com/c/device/google/tangorpro/+/25019613
which both included an active gain for headsets when the AEC is off.

As a safety precaution, and to allow easier spotting of such issues, the CL also
explicitly sets the gains to 0 dB also for the other headset tuning (before it was
just deactivating the gain application for that)

Bug: 309517292
Test: Manual test calls on Tangor using Meet.
Change-Id: I4f29cd780fa6ca52df543387546d2a4b4b161139